% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sotc_retrieve.R
\name{sotc_retrieve}
\alias{sotc_retrieve}
\title{State of the City Speeches}
\usage{
sotc_retrieve(meta = FALSE)
}
\arguments{
\item{meta}{Include or exclude speech metadata. Excluded by default.}
}
\value{
A data.frame with speech paragraphs in rows
}
\description{
This function loads State of the City speeches from Github one paragraph per row.
}
